+++ src/ape/_cli.py
@@ -7,8 +7,8 @@ import click
 import importlib_metadata as metadata
 import yaml
 
-from ape.cli import Abort, ape_cli_context
-from ape.exceptions import ApeException, handle_ape_exception
+from ape.cli import ape_cli_context
+from ape.exceptions import Abort, ApeException, handle_ape_exception
 from ape.logging import logger
 from ape.plugins import clean_plugin_name
