--- src/ape/cli/utils.py
@@ -1,5 +1,0 @@
-from ape.exceptions import Abort
-
-# TODO: Delete as part of 0.7 breaking changes
-# Kept here for backwards compatibility.
-__all__ = ["Abort"]
