+++ tests/functional/test_exceptions.py
@@ -1,8 +1,7 @@
 import re
 
 from ape.api import ReceiptAPI
-from ape.cli import Abort
-from ape.exceptions import TransactionError
+from ape.exceptions import Abort, TransactionError
 from ape_ethereum.transactions import Receipt
